@debris 2017-03-06
I'm not sure about that... on my machine I see no difference. Have you compared generated assembly?