@burdges 2017-03-06
It likely depends upon your rust version, but one should broadly expect unsafe code to grow slower in future as the memory model work proceeds.