@burdges 2017-03-30
I've pushed an update to this branch for the current master and partially addressed #9