@burdges 2017-03-30
Oh https://github.com/debris/tiny-keccak/pull/11 is probably better.