@burdges 2017-03-31
Now closing since https://github.com/debris/tiny-keccak/pull/11 is merged.